refactor(cockpit): extract dashboard sections engine (#403)
Move the 15+ ``_section_*`` helpers and their formatting utilities
out of the monolithic ``cockpit.py`` (3873 LOC) into a dedicated
``pollypm.cockpit_sections`` package with one module per section and
a shared ``base.py`` carrying the ``SectionRenderer`` Protocol plus
common primitives (``_DASHBOARD_DIVIDER_WIDTH``, ``_dashboard_divider``,
``_iso_to_dt``, ``_spark_bar``, ``_STATUS_ICONS``, …).

Package layout:

* ``base.py`` — shared constants + formatters + the ``SectionRenderer`` Protocol
* ``header.py`` / ``summary.py`` / ``velocity.py`` / ``you_need_to.py`` /
  ``in_flight.py`` / ``recent.py`` / ``activity.py`` / ``insights.py`` /
  ``downtime.py`` / ``quick_actions.py`` — one section each
* ``project_dashboard.py`` — per-project orchestrator
  (``_render_project_dashboard`` + mtime-cached ``_dashboard_project_tasks``)
* ``dashboard.py`` — global Polly dashboard (``_build_dashboard``)

``cockpit.py`` keeps its orchestration surface and re-exports the
moved names so the test suite and external callers (cockpit_ui,
cockpit_palette, plugin_builtin/core_rail_items, …) continue to work
without edits. All 83 cockpit dashboard + inbox UI tests and the
wider rail/palette/worker-roster/metrics suites still pass
(175 tests verified locally).
